+++ CHANGELOG.md
@@ -25,6 +25,7 @@ All notable changes to this project will be documented in this file.
 - **plugin-text**. Added colors plugin
 - **plugin-text**. Added button to wrap in a blockquote plugin
 - **plugin-rows**. Pass `remove` and `replace` functions to childs
+- **all**. Added plugin icons and descriptions
 
 ### Changed